+++ CHANGELOG.md
@@ -32,6 +32,10 @@ Test bad parameters release
 
 [444]: https://github.com/atomist/rug/issues/444
 
+### Added
+
+-   Query by example option for authoring path expressions against model
+
 ## [0.15.0] - 2017-03-17
 
 [0.15.0]: https://github.com/atomist/rug/compare/0.14.0...0.15.0